Fix cast

+++ i386/i386at/model_dep.c
@@ -409,9 +409,9 @@ i386at_init(void)
 #endif	/* MACH_XEN */
 #if PAE
 #ifdef __x86_64__
-	set_cr3((unsigned)_kvtophys(kernel_pmap->l4base));
+	set_cr3((unsigned long)_kvtophys(kernel_pmap->l4base));
 #else
-	set_cr3((unsigned)_kvtophys(kernel_pmap->pdpbase));
+	set_cr3((unsigned long)_kvtophys(kernel_pmap->pdpbase));
 #endif
 #ifndef	MACH_HYP
 	if (!CPU_HAS_FEATURE(CPU_FEATURE_PAE))
@@ -419,7 +419,7 @@ i386at_init(void)
 	set_cr4(get_cr4() | CR4_PAE);
 #endif	/* MACH_HYP */
 #else
-	set_cr3((unsigned)_kvtophys(kernel_page_dir));
+	set_cr3((unsigned long)_kvtophys(kernel_page_dir));
 #endif	/* PAE */
 #ifndef	MACH_HYP
 	/* already set by Hypervisor */
